pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "luister"
version = "0.1.0"
description = "A music player built with PyQt6"
authors = [
    {name = "Yilmaz Mustafa", email = "ymus@tuta.io"}
]
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Sound/Audio :: Players",
]
dependencies = [
    "click>=8.0.0",
    "PyQt6>=6.6.0",
    "python-dotenv>=0.19.0",
    "librosa>=0.10.0",
    "tinytag>=1.8.1",
    "openai-whisper>=20230314",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0.0",
    "pytest-qt>=3.0.0",
    "black>=21.0.0",
    "flake8>=4.0.0",
    "opentelemetry-sdk>=1.24.0",
    "opentelemetry-api>=1.24.0",
]

[project.scripts]
luister = "luister:main"

[tool.setuptools.packages.find]
where = ["src"]
include = ["luister*"]
exclude = ["tests*", "docs*"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
qt_api = "pyqt5"
